requirements.txt
langchain
langchain-huggingface
huggingface-hub
transformers
python-dotenv
pandas
streamlit